I now use the PHP session array to set the dummy instructor id.
Component dependencies moved outside of component file and into the demo page. I changed the JS code to grab a global JS variable which represents the current instructor's ID. This global variable is set within the head of the demo page by using PHP to echo the value from the session array.